+++ src/lib/tta_dec.c
@@ -31,8 +31,7 @@ tta_decode_mch(
 	i32 *const dest, const u8 *const, u32 *const restrict crc_out,
 	/*@out@*/ size_t *const restrict ni32_out,
 	struct BitCache *const restrict bitcache,
-	struct Codec *const restrict codec, size_t, size_t, uint, i32, uint,
-	uint
+	struct Codec *const restrict codec, size_t, size_t, u8, i32, u8, uint
 )
 /*@modifies	*dest,
 		*crc_out,
@@ -54,7 +53,7 @@ tta_decode_1ch(
 	i32 *const dest, const u8 *const, u32 *const restrict crc_out,
 	/*@out@*/ size_t *const restrict ni32_out,
 	struct BitCache *const restrict bitcache,
-	struct Codec *const restrict codec, size_t, size_t, uint, i32, uint
+	struct Codec *const restrict codec, size_t, size_t, u8, i32, u8
 )
 /*@modifies	*dest,
 		*crc_out,
@@ -76,7 +75,7 @@ tta_decode_2ch(
 	i32 *const dest, const u8 *const, u32 *const restrict crc_out,
 	/*@out@*/ size_t *const restrict ni32_out,
 	struct BitCache *const restrict bitcache,
-	struct Codec *const restrict codec, size_t, size_t, uint, i32, uint
+	struct Codec *const restrict codec, size_t, size_t, u8, i32, u8
 )
 /*@modifies	*dest,
 		*crc_out,
@@ -112,9 +111,9 @@ libttaR_tta_decode(
 @*/
 {
 	size_t r;
-	const uint predict_k = tta_predict_k(samplebytes);
+	const  u8 predict_k = tta_predict_k(samplebytes);
 	const i32 filter_round = tta_filter_round(samplebytes);
-	const uint filter_k = tta_filter_k(samplebytes);
+	const  u8 filter_k = tta_filter_k(samplebytes);
 	size_t safety_margin = (
 		src_len - (TTABUF_SAFETY_MARGIN * nchan * samplebytes)
 	);
@@ -206,7 +205,7 @@ tta_decode_mch(
 	/*@out@*/ size_t *const restrict ni32_out,
 	struct BitCache *const restrict bitcache,
 	struct Codec *const restrict codec, size_t ni32_target,
-	size_t safety_margin, uint predict_k, i32 filter_round, uint filter_k,
+	size_t safety_margin, u8 predict_k, i32 filter_round, u8 filter_k,
 	uint nchan
 )
 /*@modifies	*dest,
@@ -274,7 +273,7 @@ tta_decode_1ch(
 	/*@out@*/ size_t *const restrict ni32_out,
 	struct BitCache *const restrict bitcache,
 	struct Codec *const restrict codec, size_t ni32_target,
-	size_t safety_margin, uint predict_k, i32 filter_round, uint filter_k
+	size_t safety_margin, u8 predict_k, i32 filter_round, u8 filter_k
 )
 /*@modifies	*dest,
 		*crc_out,
@@ -324,7 +323,7 @@ tta_decode_2ch(
 	/*@out@*/ size_t *const restrict ni32_out,
 	struct BitCache *const restrict bitcache,
 	struct Codec *const restrict codec, size_t ni32_target,
-	size_t safety_margin, uint predict_k, i32 filter_round, uint filter_k
+	size_t safety_margin, u8 predict_k, i32 filter_round, u8 filter_k
 )
 /*@modifies	*dest,
 		*crc_out,
